Split `UserTagControl` out into partial classes
